fix: Update return type for model() to an Eloquent Model
By using `object` we wasn't ensuring that the method chaining would be valid.
Example `$this->model()->findOrFail($id);` my not exist, In this case we need
contract the implementation of an Eloquent Model. This may warrant changing the
classnames to include `Eloquent` in them if we want to support other implementations.